Make the gameobject more user friendly
Removed all unnessesary drag and drop parameters on the game object,
such as the "MainOrbitImproved" script needed the user to manually
assign the CameraReference object as target.

Every prefab should now work without any extra work needed.  (which is
great for the designers)

Every script is now in the same folder. This is to make coding easier
since some script couldn't make references to some of the other scripts
before. This is now possible.