Removed Dependensies of ThirdPersonController and CorssPlatformInput + Made the sunshafts not needing a camera manually assigned
Everything conserning the player prefab is now in the Player folder in
Assets/Old World/Player.

The default size of the camera array in the sunshaft script is now 0,
which will make the main camera assigned by default. If more than one
camera is used, increase the size of this array in the unity editor and
attach the cameras.