Continued work on battery room and code
Created link for battery slots that only start when both are online.
Batteries still have problems, primarily that they don't know when they
are put into a slot and when they are picked up. Will fix tomorrow.
Big change to battery chargers: The script (ChargerScript) is now placed
on the battery slot that is a child to the "charger block" (the big
black one). This to make the positioning of the battery easier to
implement.